Todos os elementos presentes na maniulação da nota fiscal que usam números já estão usando a entrada numérica e o respectivo binding já tem um tipo numérico.

+++ BibliotecaCentral/ModeloXML/PartesProcesso/PartesNFe/PartesDetalhes/Transporte.cs
@@ -7,7 +7,7 @@ namespace BibliotecaCentral.ModeloXML.PartesProcesso.PartesNFe.PartesDetalhes
     public class Transporte
     {
         [XmlElement(Order = 0)]
-        public int modFrete { get; set; }
+        public byte modFrete { get; set; }
 
         [XmlElement(Order = 1)]
         public Motorista transporta { get; set; }
